requirements.txt
torch
torchvision
numpy
faiss-cpu
matplotlib
streamlit
pillow
wikipedia